# parus/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Category": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"parent_id": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"default": null
					},
					"level": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean",
						"default": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"level",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCategoryDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"parent_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCategoryDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"parent_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryWithCount": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"parent_id": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"default": null
					},
					"level": {
						"type": "number",
						"format": "double"
					},
					"is_active": {
						"type": "boolean",
						"default": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"products_count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"level",
					"created_at",
					"updated_at",
					"products_count"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "product-catalog-api",
		"version": "1.0.0",
		"description": "Product catalog REST API with categories and subcategories",
		"contact": {}
	},
	"paths": {
...
		"/api/categories": {
			"post": {
				"operationId": "CreateCategory",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCategoryDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetActiveCategories",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CategoryWithCount"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/categories/{id}": {
			"put": {
				"operationId": "UpdateCategory",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Category"
								}
							}
						}
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCategoryDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCategory",
				"responses": {
					"200": {
						"description": ""
					}
				},
				"tags": [
					"Categories"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}